# applications/rpg_table_helper/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Main API v1.0",
    "version": "v1.0"
  },
  "paths": {
...
    "/PlayerCharacter/createcharacter": {
      "post": {
        "tags": [
          "PlayerCharacter"
        ],
        "summary": "Creates a new player character with the calling user as owner.",
        "requestBody": {
          "description": "The creation details",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerCharacterCreateDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerCharacterCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerCharacterCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerCharacterCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The id of the newly created player character",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerCharacterIdentifier"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerCharacterIdentifier"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerCharacterIdentifier"
                }
              }
            }
          },
          "400": {
            "description": "If the data provided is invalid",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "If you are not logged in",
            "content": {
              "text/plain": {
                "schema": {
                  "oneOf": [{
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              },
              "application/json": {
                "schema": {
                  "oneOf": [{
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "oneOf": [{
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/PlayerCharacter/getplayercharacters": {
      "get": {
        "tags": [
          "PlayerCharacter"
        ],
        "summary": "Returns a list of player characters for the calling user.",
        "responses": {
          "200": {
            "description": "A list of all playerCharacters for this user",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerCharacter"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerCharacter"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerCharacter"
                  }
                }
              }
            }
          },
          "400": {
            "description": "If there was an error retrieving the playerCharacters",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "If you are not logged in",
            "content": {
              "text/plain": {
                "schema": {
                  "oneOf": [{
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              },
              "application/json": {
                "schema": {
                  "oneOf": [{
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "oneOf": [{
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/PlayerCharacter/getplayercharacter/{playercharacterid}": {
      "get": {
        "tags": [
          "PlayerCharacter"
        ],
        "summary": "Returns a single playerCharacter.",
        "description": "You must be the owner of this playerCharacter",
        "parameters": [{
          "name": "playercharacterid",
          "in": "path",
          "description": "The id of the desired playerCharacter",
          "required": true,
          "schema": {
            "type": "string"
          }
        }],
        "responses": {
          "200": {
            "description": "The playerCharacter",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerCharacter"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerCharacter"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerCharacter"
                }
              }
            }
          },
          "400": {
            "description": "If there was an error retrieving the playerCharacters",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "If you are not logged in or you are not allowed to see this playerCharacter",
            "content": {
              "text/plain": {
                "schema": {
                  "oneOf": [{
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              },
              "application/json": {
                "schema": {
                  "oneOf": [{
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "oneOf": [{
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/PlayerCharacter/getplayercharactersincampagne": {
      "get": {
        "tags": [
          "PlayerCharacter"
        ],
        "summary": "Returns a list of player characters for a given campagne (if the calling user is the dm).",
        "parameters": [{
          "name": "Value",
          "in": "query",
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }],
        "responses": {
          "200": {
            "description": "A list of all playerCharacters for this user",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerCharacter"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerCharacter"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerCharacter"
                  }
                }
              }
            }
          },
          "400": {
            "description": "If there was an error retrieving the playerCharacters",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "If you are not logged in or not the dm",
            "content": {
              "text/plain": {
                "schema": {
                  "oneOf": [{
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              },
              "application/json": {
                "schema": {
                  "oneOf": [{
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              },
              "text/json": {
                "schema": {
                  "oneOf": [{
                      "$ref": "#/components/schemas/ProblemDetails"
                    },
                    {
                      "$ref": "#/components/schemas/HttpValidationProblemDetails"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CampagneIdentifier": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "HttpValidationProblemDetails": {
        "allOf": [{
            "$ref": "#/components/schemas/ProblemDetails"
          },
          {
            "type": "object",
            "properties": {
              "errors": {
                "type": "object",
                "additionalProperties": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "nullable": true
              }
            },
            "additionalProperties": {}
          }
        ]
      },
...
      "PlayerCharacter": {
        "allOf": [{
            "$ref": "#/components/schemas/PlayerCharacterIdentifierGuidNodeModelBase"
          },
          {
            "type": "object",
            "properties": {
              "rpgCharacterConfiguration": {
                "type": "string",
                "nullable": true
              },
              "characterName": {
                "type": "string",
                "nullable": true
              },
              "playerUserId": {
                "$ref": "#/components/schemas/UserIdentifier"
              },
              "campagneId": {
                "$ref": "#/components/schemas/CampagneIdentifier"
              }
            },
            "additionalProperties": false
          }
        ]
      },
      "PlayerCharacterCreateDto": {
        "required": [
          "characterName"
        ],
        "type": "object",
        "properties": {
          "rpgCharacterConfiguration": {
            "type": "string",
            "description": "This is the json serialized configuration of the character",
            "nullable": true
          },
          "characterName": {
            "minLength": 1,
            "type": "string"
          },
          "campagneId": {
            "type": "string",
            "description": "The id of the campagne for this character",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PlayerCharacterIdentifier": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "PlayerCharacterIdentifierGuidNodeModelBase": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/PlayerCharacterIdentifier"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
...
      "UserIdentifier": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter 'Bearer' [space] and then your token in the text input below.\r\n\r\nExample: \"Bearer 1safsfsdfdfd\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [{
    "Bearer": []
  }]
}